containers/dev_container/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
gitpython==3.1.40
python-dotenv==1.0.0
aiofiles==23.2.1
